+++ Changelog.md
@@ -4,6 +4,7 @@
 * Login for mentors
 * List of mentors
 * View mentors details
+* List each mentor available hours on detail page
 
 **✨ Improvements:**